% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{class}
\name{CategoriesDataFrame}
\alias{CategoriesDataFrame}
\alias{CategoriesDataFrame-class}
\alias{show,CategoriesDataFrame-method}
\alias{default}
\alias{default,CategoriesDataFrame-method}
\alias{[<-,CategoriesDataFrame-method}
\alias{[[<-,CategoriesDataFrame,ANY,ANY,ANY-method}
\alias{[[<-,CategoriesDataFrame-method}
\alias{rownames<-,CategoriesDataFrame-method}
\alias{subset,CategoriesDataFrame-method}
\title{CategoriesDataFrame objects}
\usage{
CategoriesDataFrame(...)

\S4method{show}{CategoriesDataFrame}(object)

default(x)

\S4method{default}{CategoriesDataFrame}(x)

\S4method{[}{CategoriesDataFrame}(x, i, j, ...) <- value

\S4method{[[}{CategoriesDataFrame,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{rownames}{CategoriesDataFrame}(x) <- value

\S4method{subset}{CategoriesDataFrame}(x, ...)
}
\arguments{
\item{...}{Arguments passed on to inherited methods.}

\item{object}{`CategoriesDataFrame` object used in [validObject()] and
[show()] function calls.}

\item{x}{`CategoriesDataFrame` object.}

\item{i}{With `j` (x[i, j]]), the row slice(s) / row name(s) , otherwise
(x[[i]]) the column slice / name.}

\item{j}{The column slice / name (x[i, j]).}

\item{value}{Atomic or vector assigned to `x`.}
}
\description{
Specialized [DataFrame] class with the following additional constraints to
represent ToppGene parameters to run a ToppGene [enrich()] query:

\itemize{
\item{Fixed number of rows with fixed order for each category.}
\item{Columns can only be set to the allowed numerical and set values that
are described when showing the object.}
}

The DataFrame semantics allow quickly setting multiple parameters at a time.
Unlike typical DataFrame behavior, [show()] displays all 19 rows instead of
using ellipses.
}
\examples{
library(DFplyr)
cats <- CategoriesDataFrame()
cats <-
    cats |>
    mutate(
        PValue = 0.001,
        MaxResults = case_when(
            grepl("Onto", rownames(cats)) ~ 25L,
            .default = MaxResults))
cats
}
\seealso{
[DataFrame::DataFrame()]
}
\keyword{classes}
\keyword{methods}
