% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webform.R
\name{cluster}
\alias{cluster}
\title{Return functional enrichment of genesets.}
\usage{
cluster(symbols_list, limit = NULL, progress = TRUE)
}
\arguments{
\item{symbols}{List of character vectors of genes.}
}
\value{
DataFrame with 15 columns containing the enrichment Category, ID,
    and associated data.
}
\description{
The ToppCluster API returns many [CATEGORIES] of geneset erichment.
}
\examples{
# ToppCluster website sample data functional enrichment call:
cluster(list(Kidney = GENELIST1, Liver = GENELIST2, Pancreas = GENELIST3))
}
