% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapi-toppgene.R
\name{lookup}
\alias{lookup}
\title{Return integer Entrez IDs from gene symbols, ensembl references, etc.}
\usage{
lookup(symbols, max_tries = 3L)
}
\arguments{
\item{symbols}{Character vector of genes.}

\item{max_tries}{Number of attempts passed on to httr2::req_retry.}
}
\value{
DataFrame with 4 columns: "Submitted" symbol character vector in the
    same order as the symbols input parameter, corresponding "Entrez"
    integer IDs, "OfficialSymbol" character vector, and "Description" of
    gene.
}
\description{
The ToppGene API returns many lookup differs from Bioconductor's identifier
lookup, therefore we have to use the web API instead of the typical
Bioconductor functions of [GSEABase::mapIdentfiers()],
[AnnotationDbi::mapIds()], etc.
}
\examples{
# Sample lookup call of the ToppGene API specification:
# - FLDB is an obsolete symbol for APOB.
# - APOE is the current symbol for APOE.
# - ENSG00000113196 is an ensembl gene symbol for HAND1.
# - ENSMUSG00000020287 is a mouse gene MPG.
lookup(c("FLDB", "APOE", "ENSG00000113196", "ENSMUSG00000020287"))
}
