% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webapi-pubchem.R
\name{lookup_pubchem}
\alias{lookup_pubchem}
\title{Return table of drug identifiers to PubChem CIDs.}
\usage{
lookup_pubchem(df)
}
\arguments{
\item{df}{DataFrame with 15 columns containing the enrichment Category, ID,
and associated data.}
}
\value{
DataFrame with 2 columns subset to Category == Drug containing input
    Source, ID, and output CID.
}
\description{
Many downstream drug analyses in Bioconductor make use of PubChem CIDs but
ToppGened drug identifiers require changes prior to conversion, and the
conversion itself is involved when scaling to large lists of identifiers.
}
\details{
Therefore this function submits queries to the PubChem Power User Gateway
(PUG) in parallel for each query.
}
\examples{
library(DFplyr)
cats <- CategoriesDataFrame()
cats <-
    cats |>
    mutate(
        PValue = case_when(
            grepl("Drug", rownames(cats)) ~ PValue,
            .default = 1e-100),
        MaxResults = case_when(
            grepl("Drug", rownames(cats)) ~ 1000L,
            .default = 1L))
# EGFR gene that has hits in all drug databases.
df_enrich <- toppgene::enrich(1956L, cats)
df_cid <- lookup_pubchem(df_enrich)
df_cid
}
